requirements.txt
pymongo==4.4.0
pyserial==3.5.0
streamlit==1.24.1